removing old folders and renaming files to clean up folder structure
