+++ jpo-ode-consumer-example/README.md
@@ -1,0 +1,3 @@
+# Kafka Consumer Example
+
+This project is intended to serve as a sample consumer application to subscribe to a Kafka topic of streams.
